pyproject.toml
[project]
name = "fastapi-users-slowapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi == 0.115.11",
    "uvicorn >= 0.34.0",
    "redis >= 3.5.3",
    "apscheduler >= 3.10.4",
    "uv >= 0.5.1",
    "slowapi >= 0.1.9",
    "pydantic >= 2.8.2",
    "pydantic-settings >= 2.6.1",
    "python-jose >= 3.3.0",
    "passlib >= 1.7.4",
    "bcrypt >= 3.2.0"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.1",
    "ruff>=0.11.2"]
